feat(contexts): unified stepped New-Context wizard
Replace the New-Context chooser + Run-plugin + Create-manual modals with one
guided wizard using the shared Stepper, matching the matrix / crawler wizards:
  Source → (plugin) Pick plugin → Configure → Preview & run
  Source → (manual) Details → Create
  Source → (import) → Crawlers

The Configure step keeps the friendly attribute picker (grouped real columns +
extendedAttributes keys); the Preview step auto-runs a dry-run before Run writes
anything. Removes NewContextModal, RunPluginModal, CreateManualTreeModal.
