+++ changes/editable-org-tree.md
@@ -4,4 +4,5 @@
 - Member counts (direct and total) update immediately after re-parenting, adding, or moving a node, so the numbers on every context stay correct.
 - The tree keeps your expanded/collapsed nodes after a drag-drop move instead of collapsing back to the top.
 - The "Run plugin" dialog no longer closes when you click outside it, so a stray click can't discard what you were setting up.
-- When a plugin asks for a list of attributes (e.g. how to name org-tree nodes), you now pick them from a dropdown of real attributes with a "+ Add attribute" button, instead of editing raw JSON.
+- When a plugin asks for a list of attributes (e.g. how to name org-tree nodes), you now pick them from a dropdown — grouped into the entity's own attributes and its extended attributes — with a "+ Add attribute" button, instead of editing raw JSON. Org-tree nodes can now be named by extended attributes too (e.g. a SuccessFactors department or an extension attribute).
+- Creating a new context tree is now a single guided wizard with steps (Source → Pick plugin → Configure → Preview & run), matching the new-crawler and new-matrix wizards, with a live preview before anything is written.
